% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densratio.R
\name{densratio}
\alias{densratio}
\title{Estimate Density Ratio p_nu(x)/p_de(y)}
\usage{
densratio(x, y, method = c("uLSIF", "KLIEP"), sigma = "auto",
  lambda = "auto", kernel_num = 100, fold = 5)
}
\arguments{
\item{x}{numeric vector or matrix as data from a numerator distribution p_nu(x).}

\item{y}{numeric vector or matrix as data from a denominator distribution p_de(y).}

\item{method}{"uLSIF"(default) or "KLIEP".}

\item{sigma}{positive numeric vector as a search range of Gaussian kernel bandwidth.}

\item{lambda}{positive numeric vector as a search range of regularization parameter for uLSIF.}

\item{kernel_num}{positive integer as number of kernels.}

\item{fold}{positive integer as fold number for cross validation for KLIEP.}
}
\description{
Estimate Density Ratio p_nu(x)/p_de(y)
}

