% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_kernel_Gaussian.R
\name{euclid_distance}
\alias{euclid_distance}
\title{Compute Euclid Distance}
\usage{
euclid_distance(x, y)
}
\arguments{
\item{x}{a numeric vector.}

\item{y}{a numeric vector.}
}
\value{
euclid distance
}
\description{
Compute Euclid Distance
}

