% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLIEP.R
\name{KLIEP}
\alias{KLIEP}
\title{Estimate Density Ratio p(x)/p(y) by KLIEP (Kullback-Leibler Importance Estimation Procedure)}
\usage{
KLIEP(x, y, sigma = "auto", kernel_num = 100, fold = 5)
}
\arguments{
\item{x}{numeric vector or matrix as data from numerator.}

\item{y}{numeric vector or matrix as data from denominator.}

\item{sigma}{positive scalar as Gaussian kernel bandwidth.}

\item{kernel_num}{positive integer as number of kernels.}

\item{fold}{positive integer as fold number for cross validation.}
}
\description{
Estimate Density Ratio p(x)/p(y) by KLIEP (Kullback-Leibler Importance Estimation Procedure)
}

