% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_kernel_Gaussian.R
\name{squared_euclid_distance}
\alias{squared_euclid_distance}
\title{Compute Squared Euclid Distance}
\usage{
squared_euclid_distance(x, y)
}
\arguments{
\item{x}{a numeric vector.}

\item{y}{a numeric vector.}
}
\value{
squared euclid distance
}
\description{
Compute Squared Euclid Distance
}
