% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uLSIF.R
\name{uLSIF}
\alias{uLSIF}
\title{Estimate Density Ratio p(x)/p(y) by uLSIF (unconstrained Least-Square Importance Fitting)}
\usage{
uLSIF(x, y, sigma = 10^seq(-3, 1, length.out = 9), lambda = 10^seq(-3, 1,
  length.out = 9), kernel_num = 100, fold = 0)
}
\arguments{
\item{x}{numeric vector or matrix as data from numerator.}

\item{y}{numeric vector or matrix as data from denominator.}

\item{sigma}{positive scalar as Gaussian kernel bandwidth.}

\item{lambda}{positive scalar as regularization parameter.}

\item{kernel_num}{positive integer as number of kernels.}

\item{fold}{positive integer as fold number for cross validation.}
}
\description{
Estimate Density Ratio p(x)/p(y) by uLSIF (unconstrained Least-Square Importance Fitting)
}

